fastApi/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-openai
langchain-community
transformers
accelerate
pdfplumber
pdf2image
pytesseract
Pillow
pandas
openpyxl
python-docx
langchain-teddynote
requests
